+++ recipes/client.rb
@@ -11,6 +11,12 @@ package "munin-node" do
   action :install
 end
 
+template "/etc/munin/munin-node.conf" do
+  source "munin-node.conf.erb"
+  mode "0644"
+  notifies :restart, "service[munin-node]"
+end
+
 service "munin-node" do
   action [:enable, :start]
 end